Upgrade gradio to 4.25
Latest version is 4.27, but it looks like 4.26 and 4.27 both have a problem where when I load the page, it starts logging lots and lots of requests coming in, with increasing sequential port numbers (see video shared in team chat from today). It looks like it has something to do with queuing?

We should retest that when we next upgrade gradio.

Couldn't find anything relevant in the changelog, at first sight: https://www.gradio.app/changelog