Trying to fix big font size in chatbot
This doesn't happen for me locally. It might have something to do with Markdown formatting though?

In the remote instance, I see that Gradio adds a class .prose to the .chatbot div, locally this doesn't happen for me. Adding or removing the class manually in the browser adds or removes Markdown formatting rendering.

Trying to hack this for now by forcing the font-size in .chatbot.prose to the regular font size (which will affect <p>, but not subheadings like <hN>