fix: Downgrade Docker image
The build started to fail: https://github.com/tw-haiven/haiven/actions/runs/10525075051/job/29163133097

Messages were around unicode characters in the Next build. When reproducing locally, it was failing for older versions of the code as well, versions that had green builds in the past.

Looks like the base Docker image for the node build stage is the problem. Coincidentally (?), the Dockerfile that I THINK is the basis of that Docker image was changed yesterday? https://github.com/nodejs/docker-node/blob/5313e6aa0b794521a732c9913151a13384d40c9c/22/alpine3.19/Dockerfile

The only way I found to fix this for now was to downgrade to an older image, one that also has an older Node version, so we have to temporarily downgrade to node 20 as well, until this image works again or we find a better one.