feat: Remove icons from the submenu entries
They were indicating which tasks are backed by the knowledge pack and are therefore customizable, and tasks which are "hard-coded" in Haiven. As the user themself cannot customize prompts at the moment, it's not really relevant to the user, so we're removing them. But the prompt titles on the chat views now have a little info tooltip that indicates that this is a customizable prompt.