Change vars into functions, to fix test collection error
In GH Actions, and on other team members' machines, there are test collection errors when running unit tests, but it was working on my machine. Trying to fix this by not instantiating variables directly, but putting them into functions that only get called when the test gets called.

Original issue of why it is different on my machine not explained yet