Fix style: Enable scroll on cards pages
When the cards were breaking longer than the viewport, we didn't get a scrollbar. Removing the "height: 100%" fixed it, but then something broke with the "height: 5%" of the collapsed prompt options and the disclaimer, so had to fix that as well.

Also made the disclaimer text a bit smaller, to reduce the probability that it breaks into multiple lines, which always looks ugly