Refactor feature toggles: Init in App, pass on the object
In prep for having instance feature toggles, want to do the full init of it in App - but to make this work well with tests, it's easiest to pass on the final toggle object through the components.

Tried versions with a hook as well, but that totally messed up the testing, because then it becomes a kind of global state, and the initialisation order matters, and the full toggle hook needs to be mocked somehow, felt messy.